Move airly coordinator to its own file
